requirements.txt
streamlit
tensorflow
numpy
pillow
h5py 